minus OK

+++ src/lexer.h
@@ -108,8 +108,8 @@ class Lexer {
         }
 
         // 数字を格納するnumValのgetter, setter
-        uint64_t getNumVal() { return numVal; }
-        void setnumVal(uint64_t numval) { numVal = numval; }
+        int getNumVal() { return numVal; }
+        void setnumVal(int numval) { numVal = numval; }
 
         // 識別子を格納するIdentifierStrのgetter, setter
         std::string getIdentifier() { return identifierStr; }
